Code review
Added error mapping to webpage so if someone eneter the wrong url they redirect to an error page which has a link to send them back to home page. Added comments in controller class so that once you add more requestmappings it will be easier to determine which one is for what.
P.S for some reason the maps on home page weren't showing up on my browser, i tried fixing it but couldn't, you might wanna test it on a couple different browsers, i tried on chrome.